+++ PassXYZLib/Platforms/Android/PxEnvironment.cs
@@ -3,5 +3,9 @@
     // All the code in this file is only included on Android.
     public partial class PxEnvironment
     {
+        public static partial string GetRoot()
+        {
+            return "/storage/emulated/0/Documents";
+        }
     }
 }