+++ PassXYZLib/Platforms/iOS/PxEnvironment.cs
@@ -3,5 +3,9 @@
     // All the code in this file is only included on iOS.
     public partial class PxEnvironment
     {
+        public static partial string GetRoot()
+        {
+            return Environment.GetFolderPath(Environment.SpecialFolder.LocalApplicationData);
+        }
     }
 }