+++ PassXYZLib/PxEnvironment.cs
@@ -1,7 +1,0 @@
-﻿namespace PassXYZLib
-{
-    // All the code in this file is included in all platforms.
-    public partial class PxEnvironment
-    {
-    }
-}