+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### 2.2.2Alpha
 - Updated to .NET 8
+- Added `PxEnvironment` and test project `UITest`
 
 ### 2.2.1
 - Updated PassXYZ to 3.2.0